Split off JIT part
Authored with Claude.
